+++ ReduxCore/inc/fields/button_set/field_button_set.js
@@ -22,7 +22,7 @@
 
     redux.field_objects.button_set.init = function( selector ) {
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-button_set' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-button_set:visible' );
         }
 
         $( selector ).each(
@@ -31,6 +31,9 @@
                 var parent = el;
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
+                }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
                 }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );