+++ ReduxCore/inc/fields/checkbox/field_checkbox.js
@@ -21,7 +21,7 @@
 
     redux.field_objects.checkbox.init = function( selector ) {
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-checkbox' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-checkbox:visible' );
         }
 
         $( selector ).each(
@@ -31,19 +31,21 @@
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
                 }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
+                }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );
                 } else {
                     return;
                 }
-
                 el.find( '.checkbox' ).on(
                     'click', function( e ) {
                         var val = 0;
-                        if ($(this).is(':checked')) {
-                            val = $(this).parent().find('.checkbox-check' ).attr('data-val');
+                        if ( $( this ).is( ':checked' ) ) {
+                            val = $( this ).parent().find( '.checkbox-check' ).attr( 'data-val' );
                         }
-                        $(this).parent().find('.checkbox-check').val(val);
+                        $( this ).parent().find( '.checkbox-check' ).val( val );
                         redux_change( $( this ) );
                     }
                 );