+++ ReduxCore/inc/fields/color_gradient/field_color_gradient.js
@@ -21,7 +21,7 @@
     redux.field_objects.color_gradient.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-color_gradient' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-color_gradient:visible' );
         }
 
         $( selector ).each(
@@ -32,7 +32,9 @@
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
                 }
-                
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
+                }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );
                 } else {