+++ ReduxCore/inc/fields/color_rgba/field_color_rgba.js
@@ -12,6 +12,7 @@
     );
 
     $.reduxColorRGBA.color_rgba = function() {
+
         $( '.redux-color_rgba-init' ).minicolors(
             {
                 animationSpeed: 50,