+++ ReduxCore/inc/fields/date/field_date.js
@@ -14,7 +14,7 @@
 
     redux.field_objects.date.init = function( selector ) {
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-date' );
+            selector = $( document ).find( '.redux-container-date:visible' );
         }
         $( selector ).each(
             function() {
@@ -22,6 +22,9 @@
                 var parent = el;
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
+                }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
                 }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );