+++ ReduxCore/inc/fields/gallery/field_gallery.js
@@ -18,7 +18,7 @@
 
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-gallery' );
+            selector = $( document ).find( '.redux-container-gallery:visible' );
         }
 
         $( selector ).each(
@@ -27,6 +27,9 @@
                 var parent = el;
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
+                }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
                 }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );