+++ ReduxCore/inc/fields/link_color/field_link_color.js
@@ -19,7 +19,7 @@
     redux.field_objects.link_color.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-link_color' );
+            selector = $( document ).find( '.redux-container-link_color:visible' );
         }
 
         $( selector ).each(
@@ -30,7 +30,9 @@
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
                 }
-                
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
+                }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );
                 } else {