+++ ReduxCore/inc/fields/multi_text/field_multi_text.js
@@ -15,7 +15,7 @@
     redux.field_objects.multi_text.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-multi_text' );
+            selector = $( document ).find( '.redux-container-multi_text:visible' );
         }
 
         $( selector ).each(
@@ -24,6 +24,9 @@
                 var parent = el;
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
+                }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
                 }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );