+++ ReduxCore/inc/fields/select_image/field_select_image.js
@@ -15,7 +15,7 @@
     redux.field_objects.select_image.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-select_image' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-select_image:visible' );
         }
 
         $( selector ).each(
@@ -24,6 +24,9 @@
                 var parent = el;
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
+                }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
                 }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );