+++ ReduxCore/inc/fields/slider/field_slider.js
@@ -15,7 +15,7 @@
     redux.field_objects.slider.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-slider' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-slider:visible' );
         }
 
         $( selector ).each(
@@ -26,7 +26,9 @@
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
                 }
-                
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
+                }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );
                 } else {