+++ ReduxCore/inc/fields/slides/field_slides.js
@@ -15,7 +15,7 @@
     redux.field_objects.slides.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-slides' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-slides:visible' );
         }
 
         $( selector ).each(
@@ -27,6 +27,9 @@
                 var parent = el;
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
+                }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
                 }
                 
                 if ( parent.hasClass( 'redux-container-slides' ) ) {