+++ ReduxCore/inc/fields/sortable/field_sortable.js
@@ -17,7 +17,7 @@
     redux.field_objects.sortable.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-sortable' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-sortable:visible' );
         }
 
         $( selector ).each(
@@ -26,6 +26,9 @@
                 var parent = el;
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
+                }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
                 }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );