+++ ReduxCore/inc/fields/sorter/field_sorter.js
@@ -23,7 +23,7 @@
     redux.field_objects.sorter.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-sorter' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-sorter:visible' );
         }
 
         $( selector ).each(
@@ -32,6 +32,9 @@
                 var parent = el;
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
+                }
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
                 }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );