+++ ReduxCore/inc/fields/typography/field_typography.js
@@ -34,7 +34,7 @@
     redux.field_objects.typography.init = function( selector ) {
 
         if ( !selector ) {
-            selector = $( document ).find( '.redux-container-typography' );
+            selector = $( document ).find( ".redux-group-tab:visible" ).find( '.redux-container-typography:visible' );
         }
 
         $( selector ).each(
@@ -45,7 +45,9 @@
                 if ( !el.hasClass( 'redux-field-container' ) ) {
                     parent = el.parents( '.redux-field-container:first' );
                 }
-                
+                if ( parent.is( ":hidden" ) ) { // Skip hidden fields
+                    return;
+                }
                 if ( parent.hasClass( 'redux-field-init' ) ) {
                     parent.removeClass( 'redux-field-init' );
                 } else {